refactor: Extract test method into _test file
