+++ plugins/source/azure/resources/services/compute/virtual_machine_scale_set_vm_network_interfaces.go
@@ -2,16 +2,12 @@ package compute
 
 import (
 	"context"
-	"encoding/json"
-	"net/http"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/compute/armcompute/v4"
 	"github.com/Azure/azure-sdk-for-go/sdk/resourcemanager/network/armnetwork/v2"
 	"github.com/cloudquery/cloudquery/plugins/source/azure/client"
-	"github.com/cloudquery/plugin-sdk/faker"
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugin-sdk/transformers"
-	"github.com/gorilla/mux"
 )
 
 func virtualMachineScaleSetsNetworkInterfaces() *schema.Table {
@@ -44,29 +40,5 @@ func fetchVirtualMachineScaleSetsNetworkInterfaces(ctx context.Context, meta sch
 		}
 		res <- p.Value
 	}
-	return nil
-}
-
-func createVirtualMachineScaleSetsNetworkInterfaces(router *mux.Router) error {
-	var item armnetwork.InterfacesClientListVirtualMachineScaleSetNetworkInterfacesResponse
-	if err := faker.FakeObject(&item); err != nil {
-		return err
-	}
-
-	emptyStr := ""
-	item.NextLink = &emptyStr
-
-	router.HandleFunc("/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/microsoft.Compute/virtualMachineScaleSets/{virtualMachineScaleSetName}/networkInterfaces", func(w http.ResponseWriter, r *http.Request) {
-		b, err := json.Marshal(&item)
-		if err != nil {
-			http.Error(w, "unable to marshal request: "+err.Error(), http.StatusBadRequest)
-			return
-		}
-		if _, err := w.Write(b); err != nil {
-			http.Error(w, "failed to write", http.StatusBadRequest)
-			return
-		}
-	})
-
 	return nil
 }
